Input的类型统一为uint64_t，Output为int64_t，有效态大小为63bit。所有有关时间累计的量都统一为unsigned long类型。

+++ include/gimcs/StateInfoParser.h
@@ -41,9 +41,9 @@ class StateInfoParser
     private:
         const State_Info_Header *my_sthd;
         unsigned char *atp;
-        int act_index;
+        unsigned long act_index;
         unsigned char *eap;
-        int eat_index;
+        unsigned long eat_index;
 };
 
 } /* namespace gimcs */
